+++ packages/core/src/di/index.ts
@@ -17,8 +17,8 @@ export {InjectFlags} from './interface/injector';
 export {ɵɵdefineInjectable, defineInjectable, ɵɵdefineInjector, InjectableType, InjectorType} from './interface/defs';
 export {forwardRef, resolveForwardRef, ForwardRefFn} from './forward_ref';
 export {Injectable, InjectableDecorator, InjectableProvider} from './injectable';
-export {INJECTOR, Injector} from './injector';
-export {ɵɵinject, inject} from './injector_compatibility';
+export {Injector} from './injector';
+export {ɵɵinject, inject, INJECTOR} from './injector_compatibility';
 export {ReflectiveInjector} from './reflective_injector';
 export {StaticProvider, ValueProvider, ConstructorSansProvider, ExistingProvider, FactoryProvider, Provider, TypeProvider, ClassProvider} from './interface/provider';
 export {ResolvedReflectiveFactory, ResolvedReflectiveProvider} from './reflective_provider';
