+++ packages/core/src/render3/errors.ts
@@ -6,7 +6,11 @@
  * Use of this source code is governed by an MIT-style license that can be
  * found in the LICENSE file at https://angular.io/license
  */
+import {InjectorType} from '../di/interface/defs';
+import {stringify} from '../util/stringify';
+
 import {TNode} from './interfaces/node';
+
 
 /** Called when directives inject each other (creating a circular dependency) */
 export function throwCyclicDependencyError(token: any): never {
@@ -30,4 +34,21 @@ export function throwErrorIfNoChangesMode(
   }
   // TODO: include debug context
   throw new Error(msg);
+}
+
+export function throwMixedMultiProviderError() {
+  throw new Error(`Cannot mix multi providers and regular providers`);
+}
+
+export function throwInvalidProviderError(
+    ngModuleType?: InjectorType<any>, providers?: any[], provider?: any) {
+  let ngModuleDetail = '';
+  if (ngModuleType && providers) {
+    const providerDetail = providers.map(v => v == provider ? '?' + provider + '?' : '...');
+    ngModuleDetail =
+        ` - only instances of Provider and Type are allowed, got: [${providerDetail.join(', ')}]`;
+  }
+
+  throw new Error(
+      `Invalid provider for the NgModule '${stringify(ngModuleType)}'` + ngModuleDetail);
 }
