+++ packages/core/test/bundling/injection/bundle.golden_symbols.json
@@ -6,19 +6,10 @@
     "name": "CIRCULAR"
   },
   {
-    "name": "CIRCULAR"
-  },
-  {
-    "name": "EMPTY"
-  },
-  {
     "name": "EMPTY_ARRAY"
   },
   {
     "name": "EmptyErrorImpl"
-  },
-  {
-    "name": "IDENT"
   },
   {
     "name": "INJECTOR"
@@ -31,12 +22,6 @@
   },
   {
     "name": "InjectionToken"
-  },
-  {
-    "name": "Injector"
-  },
-  {
-    "name": "MULTI_PROVIDER_FN"
   },
   {
     "name": "NEW_LINE"
@@ -58,9 +43,6 @@
   },
   {
     "name": "NO_NEW_LINE"
-  },
-  {
-    "name": "NULL_INJECTOR"
   },
   {
     "name": "NULL_INJECTOR"
@@ -93,16 +75,13 @@
     "name": "SkipSelf"
   },
   {
-    "name": "StaticInjector"
+    "name": "THROW_IF_NOT_FOUND"
   },
   {
     "name": "USE_VALUE"
   },
   {
     "name": "UnsubscriptionErrorImpl"
-  },
-  {
-    "name": "_THROW_IF_NOT_FOUND"
   },
   {
     "name": "__forward_ref__"
@@ -121,9 +100,6 @@
   },
   {
     "name": "catchInjectorError"
-  },
-  {
-    "name": "computeDeps"
   },
   {
     "name": "couldBeInjectableType"
@@ -195,37 +171,28 @@
     "name": "makeRecord"
   },
   {
-    "name": "multiProviderMixError"
-  },
-  {
     "name": "providerToFactory"
   },
   {
     "name": "providerToRecord"
   },
   {
-    "name": "recursivelyProcessProviders"
-  },
-  {
     "name": "resolveForwardRef"
-  },
-  {
-    "name": "resolveProvider"
-  },
-  {
-    "name": "resolveToken"
   },
   {
     "name": "setCurrentInjector"
   },
   {
-    "name": "staticError"
-  },
-  {
     "name": "stringify"
   },
   {
-    "name": "tryResolveToken"
+    "name": "throwCyclicDependencyError"
+  },
+  {
+    "name": "throwInvalidProviderError"
+  },
+  {
+    "name": "throwMixedMultiProviderError"
   },
   {
     "name": "ɵɵdefineInjectable"
